+++ CHANGELOG.md
@@ -1,16 +1,21 @@
 # Change Log
-## [Unreleased]
+## [3.X.X - Unreleased]
 ### Added
 - Predefined support for GitLab.
+- Predefined support for Gogs.
+
+### Changed
+- Moved a number of models from `useintest.models` to `useintest.services.models`.
+
 
 ## 2.0.0 - 2017-01-11
 ### Changed
 - Switched to using native iRODS controllers, executables and helpers (no longer requires `test-with-irods`).
 - Groups predefined controllers by product (e.g. iRODS) not by type (e.g. service).
 - Allows quicker testing with latest versions of controllers only.
-- Rebranded from `startfortest` to `useintest`.
+- Re-branded from `startfortest` to `useintest`.
 - Moved documentation to ReadTheDocs.
 
 ## 1.0.0 - 2016-11-23
 ### Added
-- First stable release
+- First stable release.